YUK-7870 Bulk Import in Bulk Operations creates devices with 5 min LP rate rather than 1 hour rate.  - removed overriding of lp interval for 400 series in the new bulk code.  - removed overriding of lp interval for 400 series meters in dbeditor code.  Changed the default lp interval value from 5m to 1h and then made dbeditor override the 2xx and 3xx meters to have 5m to keep legacy support.

+++ yukon-client/common/src/com/cannontech/database/db/device/DeviceLoadProfile.java
@@ -7,7 +7,7 @@ public class DeviceLoadProfile extends com.cannontech.database.db.DBPersistent
 {
 	private Integer deviceID = null;
 	private Integer lastIntervalDemandRate = new Integer(300);
-	private Integer loadProfileDemandRate = new Integer(300);
+	private Integer loadProfileDemandRate = new Integer(3600);
 	private String loadProfileCollection = "NNNN";
 	private Integer voltageDmdInterval = new Integer(60);
 	private Integer voltageDmdRate = new Integer(300);
